pyproject.toml
[project]
name = "tea-search"
version = "0.1.0"
description = "Multi-agent TEA search pipeline for technique enrichment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "arxiv>=2.1.0",
    "google-api-python-client>=2.100.0",
    "requests>=2.31.0",
    "claude-agent-sdk>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["test", "lib/tea-search/tests"]
pythonpath = ["."]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]